requirements.txt
sentence-transformers==2.2.2
huggingface-hub==0.14.1
langchain
flask
pypdf
python-dotenv
pinecone-client[grpc]
langchain-pinecone
langchain-community
langchain-openai
langchain-experimental




